src/requirements.txt
Flask==2.2.3
Werkzeug==2.2.2
Flask-Parameter-Validation==2.1.3
Flask-Authorize
Flask-Babel == 2.0.0
PyJWT
serverless-wsgi
sqlalchemy-pagination
boto3
aws-psycopg2
SQLAlchemy
coverage
pytest
HTMLParser
flask-swagger
requests
psycopg2-binary==2.9.9
